Enhance feed listing components with new empty state handling
- Introduced `Empty` and `FilterEmpty` components to manage empty states in the feed listing, improving user experience when no listings are available.
- Refactored `Content` and `ListingList` components to utilize the new empty state components, streamlining the rendering logic.
- Updated translations in `cs.yaml` and `en.yaml` for new empty state messages, ensuring clarity for users.
- Removed outdated `Empty` component from `ListingList` to reduce redundancy and improve maintainability.
