+++ .vscode/settings.json
@@ -187,6 +187,7 @@
 		"exploitům",
 		"Facebooku",
 		"fajn",
+		"famicons",
 		"farmení",
 		"farmit",
 		"farmitelné",
