+++ apps/app/src/app/@buyer/feed/FeedListingPage/ListingList/ListingList.tsx
@@ -7,13 +7,10 @@ import {
 	VisibleContainer,
 	SpinnerContainer as VisibleSpinnerContainer,
 } from "@use-pico/client/ui/container";
-import { EmptyState } from "@use-pico/client/ui/empty-state";
 import { withFallback } from "@use-pico/client/utils";
 import type { tListingQuery } from "@zbav-se.me/sdk/api/buyer";
 import { withListingQuery } from "@zbav-se.me/sdk/query/buyer/listing";
-import { type ReactNode, Suspense, useCallback, useEffect, useMemo, useRef } from "react";
-import { Empty } from "./Empty";
-import { FilterEmpty } from "./FilterEmpty";
+import { type ReactNode, Suspense, useCallback, useEffect, useRef } from "react";
 import { Item } from "./Item";
 
 export namespace ListingList {
@@ -60,29 +57,6 @@ export const ListingList = withFallback(
 		});
 
 		const { data: listingCollection } = withListingQuery.useCollectionQuery(query);
-		const { data: listingCount } = withListingQuery.useCountQuery(query);
-		const check = useMemo(() => {
-			return [
-				{
-					check() {
-						return listingCount.isEmpty;
-					},
-					render() {
-						return <Empty />;
-					},
-				},
-				{
-					check() {
-						return listingCount.isFilterEmpty;
-					},
-					render() {
-						return <FilterEmpty />;
-					},
-				},
-			] satisfies EmptyState.Check[];
-		}, [
-			listingCount,
-		]);
 
 		const placeholder = useCallback(() => {
 			return <VisibleSpinnerContainer />;
@@ -100,32 +74,30 @@ export const ListingList = withFallback(
 				}}
 				{...props}
 			>
-				<EmptyState check={check}>
-					<VisibilityProvider store={visibility}>
-						{listingCollection.map((listingId) => (
-							<VisibleContainer
-								key={listingId}
-								id={listingId}
-								data-id={listingId}
-								placeholder={placeholder}
-								ui={{
-									height: "full",
-									width: "full",
-								}}
-							>
-								<Suspense fallback={<Item.Fallback />}>
-									<Item
-										listingId={listingId}
-										feedId={feedId}
-										withScore={withScore}
-									/>
-								</Suspense>
-							</VisibleContainer>
-						))}
+				<VisibilityProvider store={visibility}>
+					{listingCollection.map((listingId) => (
+						<VisibleContainer
+							key={listingId}
+							id={listingId}
+							data-id={listingId}
+							placeholder={placeholder}
+							ui={{
+								height: "full",
+								width: "full",
+							}}
+						>
+							<Suspense fallback={<Item.Fallback />}>
+								<Item
+									listingId={listingId}
+									feedId={feedId}
+									withScore={withScore}
+								/>
+							</Suspense>
+						</VisibleContainer>
+					))}
 
-						{appendix}
-					</VisibilityProvider>
-				</EmptyState>
+					{appendix}
+				</VisibilityProvider>
 			</Container>
 		);
 	},
