+++ apps/app/src/translation/cs.yaml
@@ -22,9 +22,15 @@
 0486f4d74790451ea29d5079202aeab5bfe5039aeaee79499694bd93:
     ref: Listing category (title)
     value: Kategorie
+04cf9e0686a93465f2354dfd8553727d51dd7516e95953b741620240:
+    ref: Go to drafts (link)
+    value: Nabídni něco k prodeji
 050d325dc0ab4741ff8dd2e315e563bf0a746ea997e21d31bc54f8dd:
     ref: Listing category not selected
     value: Kategorii nemáš vybranou
+053186ab8ac076e257afc816ae62151a4fb7bc87e4e1927dc5312e2b:
+    ref: No listings - all empty (message)
+    value: Máš tu čest vystavit úplně první inzerát a vyzkoušet, jak systém funguje!
 056c5d8421348af4114fe316b813803bc6d3a9f939e93b4487b32779:
     ref: Listing expiration (hint)
     value: |
@@ -441,6 +447,9 @@
 61324250f05e5b53c0994000ad235a587fa6c894a2414d561a434548:
     ref: Draft list (title)
     value: Rozpracované inzeráty
+615b016adf6d7fbbaeadbba5cef67b63930f5eea2c5729100da00315:
+    ref: No listings - all empty (title)
+    value: Jsi úplně první!
 61dd087b73e6d2b3291aa14c76d9dcc813946be2957f788836796746:
     ref: Transaction status closed (label)
     value: Obchod uzavřen
@@ -560,9 +569,6 @@
 847f52b94d2fd82f89cf548490067c305182a98ca1342616d166ae6a:
     ref: Personal - Phone (placeholder)
     value: Tvoje telefonní číslo
-85ac6be17be3adb14aa92cb457175f48d0dd61a23f5766a50c160d51:
-    ref: No listings (title)
-    value: Kde nic, tu nic. Ale všechno bude, všechno bude!
 86856c2ffab9a4998ded4e5585299bf873d60db1bcaed1e4f8f13869:
     ref: Image not available anymore
     value: Obrázek už není dostupný, tohle by se nemělo stát.
